refactoring: 소셜 로그인 리팩토링 완성
