+++ src/main/java/com/sporthustle/hustle/oauth/OAuthService.java
@@ -1,15 +1,17 @@
 package com.sporthustle.hustle.oauth;
 
+import static com.sporthustle.hustle.common.consts.Constants.*;
+
 import com.google.gson.JsonElement;
 import com.google.gson.JsonParser;
 import com.sporthustle.hustle.common.exception.BaseException;
 import com.sporthustle.hustle.common.jwt.dto.KakaoTokenInfo;
 import com.sporthustle.hustle.oauth.dto.OAuthUserInfoResponseDTO;
+import com.sporthustle.hustle.oauth.utils.OAuthConnector;
+import com.sporthustle.hustle.oauth.utils.kakao.KakaoInformationParser;
 import java.io.*;
 import java.net.HttpURLConnection;
 import java.net.URL;
-
-import com.sporthustle.hustle.oauth.utils.KakaoInformationParser;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.RandomStringUtils;
@@ -22,6 +24,7 @@ import org.springframework.stereotype.Service;
 public class OAuthService {
 
   private final KakaoInformationParser kakaoInformationParser;
+  private final OAuthConnector connector;
 
   @Value("${oauth.secret.client-id}")
   private String clientId;
@@ -32,49 +35,28 @@ public class OAuthService {
   public KakaoTokenInfo getKakaoToken(String code) {
     String access_Token = "";
     String refresh_Token = "";
-    String reqURL = "https://kauth.kakao.com/oauth/token";
+    String reqURL = KAKAO_REQUEST_URL_FOR_TOKEN;
 
     try {
       URL url = new URL(reqURL);
-      HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-
-      // POST 요청을 위해 기본값이 false인 setDoOutput을 true로
-      conn.setRequestMethod("POST");
-      conn.setDoOutput(true);
+      HttpURLConnection conn = connector.getConnectionForKakaoToken(url);
 
       // POST 요청에 필요로 요구하는 파라미터 스트림을 통해 전송
-      BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(conn.getOutputStream()));
-      StringBuilder sb = new StringBuilder();
-      sb.append("grant_type=authorization_code");
-      sb.append("&client_id=" + clientId); // TODO REST_API_KEY 입력
-      sb.append("&redirect_uri=" + redirectUrl); // TODO 인가코드 받은 redirect_uri 입력
-      sb.append("&code=" + code);
-      bw.write(sb.toString());
-      bw.flush();
+      connector.transportKakaoUrl(conn, clientId, redirectUrl, code);
 
       // 결과 코드가 200이라면 성공
-      int responseCode = conn.getResponseCode();
-      System.out.println("responseCode : " + responseCode);
+      printLogForConnection(conn);
 
       // 요청을 통해 얻은 JSON타입의 Response 메세지 읽어오기
-      BufferedReader br = new BufferedReader(new InputStreamReader(conn.getInputStream()));
-      String line = "";
-      String result = "";
-
-      while ((line = br.readLine()) != null) {
-        result += line;
-      }
-      System.out.println("response body : " + result);
+      String jsonResultForToken = kakaoInformationParser.getJsonForKakaoToken(conn);
 
       // Gson 라이브러리에 포함된 클래스로 JSON파싱 객체 생성
       JsonParser parser = new JsonParser();
-      JsonElement element = parser.parse(result);
+      JsonElement element = parser.parse(jsonResultForToken);
 
-      access_Token = element.getAsJsonObject().get("access_token").getAsString();
-      refresh_Token = element.getAsJsonObject().get("refresh_token").getAsString();
+      access_Token = kakaoInformationParser.getAccessToken(element);
+      refresh_Token = kakaoInformationParser.getRefreshToken(element);
 
-      br.close();
-      bw.close();
     } catch (IOException e) {
       e.printStackTrace();
     }
@@ -84,40 +66,28 @@ public class OAuthService {
 
   public OAuthUserInfoResponseDTO getKakaoUserInfo(String token) throws BaseException {
 
-    String reqURL = "https://kapi.kakao.com/v2/user/me";
+    String reqURL = KAKAO_REQUEST_URL_FOR_INFORMATION;
 
     // access_token을 이용하여 사용자 정보 조회
     try {
       URL url = new URL(reqURL);
-      HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-
-      conn.setRequestMethod("POST");
-      conn.setDoOutput(true);
-      conn.setRequestProperty("Authorization", "Bearer " + token); // 전송할 header 작성, access_token전송
+      HttpURLConnection conn = connector.getConnectionForUserInformation(url, token);
 
       // 결과 코드가 200이라면 성공
-      int responseCode = conn.getResponseCode();
-      log.info("responseCode : " + responseCode);
+      printLogForConnection(conn);
 
       // 요청을 통해 얻은 JSON타입의 Response 메세지 읽어오기
-      BufferedReader br = new BufferedReader(new InputStreamReader(conn.getInputStream()));
-      String line = "";
-      String result = "";
-
-      while ((line = br.readLine()) != null) {
-        result += line;
-      }
+      String jsonResult = kakaoInformationParser.getJsonForKakaoUserInformation(conn);
 
       // Gson 라이브러리로 JSON파싱
       JsonParser parser = new JsonParser();
-      JsonElement element = parser.parse(result);
+      JsonElement element = parser.parse(jsonResult);
 
       String snsId = kakaoInformationParser.getSnsIdByParsing(element);
       String email = kakaoInformationParser.getEmailByParsing(element);
       String name = kakaoInformationParser.getNameByParsing(element);
       String gender = kakaoInformationParser.getGenderByParsing(element);
 
-      br.close();
       return OAuthUserInfoResponseDTO.builder()
           .oauthId(snsId)
           .email(email)
@@ -132,67 +102,12 @@ public class OAuthService {
     return null;
   }
 
-//  private String getSnsIdByParsing(JsonElement element) {
-//    long id = element.getAsJsonObject().get("id").getAsLong();
-//    return Long.toString(id);
-//  }
-//
-//  private String getEmailByParsing(JsonElement element) {
-//    boolean hasEmail =
-//            element
-//                    .getAsJsonObject()
-//                    .get("kakao_account")
-//                    .getAsJsonObject()
-//                    .get("has_email")
-//                    .getAsBoolean();
-//    String email = "";
-//    if (hasEmail) {
-//      email =
-//              element
-//                      .getAsJsonObject()
-//                      .get("kakao_account")
-//                      .getAsJsonObject()
-//                      .get("email")
-//                      .getAsString();
-//    }
-//    return email;
-//  }
-//
-//  private String getNameByParsing(JsonElement element) {
-//    String name =
-//            element
-//                    .getAsJsonObject()
-//                    .get("kakao_account")
-//                    .getAsJsonObject()
-//                    .get("profile")
-//                    .getAsJsonObject()
-//                    .get("nickname")
-//                    .getAsString();
-//    return name;
-//  }
-//
-//  private String getGenderByParsing(JsonElement element) {
-//    boolean has_gender =
-//            element
-//                    .getAsJsonObject()
-//                    .get("kakao_account")
-//                    .getAsJsonObject()
-//                    .get("has_gender")
-//                    .getAsBoolean();
-//    String gender = "";
-//    if (has_gender) {
-//      gender =
-//              element
-//                      .getAsJsonObject()
-//                      .get("kakao_account")
-//                      .getAsJsonObject()
-//                      .get("gender")
-//                      .getAsString();
-//    }
-//    return gender;
-//  }
-
   private String getRamdomPassword() {
     return RandomStringUtils.randomAlphanumeric(100);
+  }
+
+  private void printLogForConnection(HttpURLConnection conn) throws IOException {
+    int responseCode = conn.getResponseCode();
+    log.info("responseCode : " + responseCode);
   }
 }
